requirements.txt
ace_tools==0.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioshutil==1.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
comtypes==1.2.0
contourpy==1.2.1
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==2.2.0
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
Flask==3.1.0
flask-sock==0.7.0
fonttools==4.53.1
forallpeople==2.6.7
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.6.0
ftfy==6.2.0
googletrans==4.0.0rc1
grpcio==1.64.1
h11==0.9.0
h2==3.2.0
handcalcs==1.6.5
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.3
hyperframe==5.2.0
idna==2.10
ifcopenshell==0.7.10
importlib_metadata==8.6.1
innerscope==0.6.0
intel-openmp==2021.4.0
ipykernel==6.29.2
ipython==8.21.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.5
jupyterlab-mathjax3==4.3.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
lark==1.1.9
litellm==1.60.8
lxml==5.3.0
lxml_html_clean==0.4.2
marker-pdf==0.2.13
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.2.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.61.1
opencv-python==4.10.0.82
openpyxl==3.1.5
openstaad==0.0.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandoc==2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.3
pdf2image==1.17.0
pdftext==0.3.10
peewee==3.17.9
pillow==10.3.0
platformdirs==4.2.0
plumbum==1.8.2
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
psutil==5.9.8
pure-eval==0.2.2
py-zerox==0.0.7
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
pyee==11.1.1
Pygments==2.17.2
PyMuPDF==1.25.3
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.3
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.34.2
Send2Trash==1.8.2
sgmllib3k==1.0.0
shapely==2.0.5
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
surya-ocr==0.4.12
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
terminado==0.18.0
texify==0.1.9
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.41.2
trio==0.30.0
trio-websocket==0.12.2
twstock==1.4.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.14.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.5.0
w3lib==2.3.1
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.55
zipp==3.21.0
python-dotenv
gunicorn  # 非常重要！這是我們在雲端要用的 Production Server
